requirements.txt
fastapi
uvicorn
python-docx
pymupdf
pandas
openpyxl
requests
python-multipart